requirements.txt
gradio
langchain
langchain-community
langchain-groq
langchain-huggingface
sentence-transformers
faiss-cpu
python-dotenv
pypdf
